added "Subscribe" tool to the editor
