+++ package.json
@@ -33,6 +33,7 @@
     "@editorjs/link": "^2.5.0",
     "@editorjs/list": "^1.8.0",
     "@editorjs/marker": "^1.3.0",
+    "@editorjs/nested-list": "^1.3.0",
     "@editorjs/paragraph": "^2.9.0",
     "@editorjs/quote": "^2.5.0",
     "@editorjs/raw": "^2.4.0",
