+++ src/components/Providers/index.tsx
@@ -4,6 +4,7 @@ import * as React from 'react'
 
 import { useApollo } from '~/lib/apollo'
 
+import { SubscribeDialog } from '../Dialog'
 import { FathomProvider } from './Fathom'
 import { GlobalNavigationContextProvider } from './GlobalNavigation'
 import { GlobalSiteContextProvider } from './GlobalSite'
@@ -25,6 +26,7 @@ export function Providers({ children, pageProps }: Props) {
 
         <GlobalNavigationContextProvider pageProps={pageProps}>
           {children}
+          <SubscribeDialog />
         </GlobalNavigationContextProvider>
       </GlobalSiteContextProvider>
     </ApolloProvider>
